split global constnts to distinguish frontend and backend env
