+++ configs/jest.cucumber.config.js
@@ -1,6 +1,6 @@
 const { jestAliases } = require('./aliases.config')
 const { getHerkinConfig } = require('./getHerkinConfig')
-const { TEST_UTILS_PATH, HERKIN_ROOT } = require('../constants')
+const { TEST_UTILS_PATH, HERKIN_ROOT } = require('../constants/backend')
 const { uniqArr } = require('@keg-hub/jsutils')
 const path = require('path')
 const glob = require('glob')
