+++ constants/frontend.js
@@ -1,0 +1,4 @@
+/**
+ * Constants for web (e.g. repos/tap)
+ */
+module.exports = {}
